third_party/blink/web_tests/http/tests/security/isolatedWorld/isolated-world-send-beacon-csp-expected.txt

CONSOLE MESSAGE: Testing main world. Request should be blocked by main world CSP.
CONSOLE ERROR: Connecting to 'http://localhost:8000/security/isolatedWorld/resources/empty.html' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE MESSAGE: Testing isolated world with no csp. Request should be blocked by main world CSP.
CONSOLE ERROR: Connecting to 'http://localhost:8000/security/isolatedWorld/resources/empty.html' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE MESSAGE: Testing isolated world with permissive csp.
PingLoader dispatched to 'http://localhost:8000/security/isolatedWorld/resources/empty.html'.
CONSOLE MESSAGE: Testing isolated world with strict csp.
CONSOLE ERROR: Connecting to 'http://localhost:8000/security/isolatedWorld/resources/empty.html' violates the following Content Security Policy directive: "connect-src 'self'". The action has been blocked.
This tests the interaction of sendBeacon command run in the isolated world with the isolated world CSP.
